# openapi-generator/aspnetcore/common/src/Agl.Cdr.Common/Controllers/CommonApi.cs
/*
 * CDR Common API
 *
 * Consumer Data Standards APIs created by the Data Standards Body (DSB), with the Data Standards Chair as the decision maker to meet the needs of the Consumer Data Right
 *
 * The version of the OpenAPI document: 1.15.0
 * Contact: contact@consumerdatastandards.gov.au
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using Agl.Cdr.Common.Attributes;
using Agl.Cdr.Common.Models;

namespace Agl.Cdr.Common.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class CommonApiController : ControllerBase
    { 
        /// <summary>
        /// Get Customer
        /// </summary>
        /// <remarks>Obtain basic information on the customer that has authorised the current session  ### Conventions In the customer payloads relevant conventions are explained here, in one place.  #### Given Names  &#x60;firstName&#x60; represents the first of a person&#39;s given names.  &#x60;middleNames&#x60; represents a collection of given names if the person has more than one given name.  Where a data holder holds a person&#39;s given names as a single string in source systems, it may not possible in some situations to reliably split these given names into their component first and middle names. In these situations, data holders MAY use the &#x60;firstName&#x60; field to return the single string of given names and an empty &#x60;middleNames&#x60; array.  For example, a person whose given names are \&quot;John Paul Winston\&quot; but the data holder cannot determine what is the first name, can return &#x60;\&quot;firstName\&quot;: \&quot;John Paul Winston\&quot;&#x60;.</remarks>
        /// <param name="xV">Version of the API end point requested by the client. Must be set to a positive integer. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If the value of [x-min-v](#request-headers) is equal to or higher than the value of [x-v](#request-headers) then the [x-min-v](#request-headers) header should be treated as absent. If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable. See [HTTP Headers](#request-headers)</param>
        /// <param name="xMinV">Minimum version of the API end point requested by the client. Must be set to a positive integer if provided. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable.</param>
        /// <param name="xFapiInteractionId">An [RFC4122](https://tools.ietf.org/html/rfc4122) UUID used as a correlation id. If provided, the data holder must play back this value in the x-fapi-interaction-id response header. If not provided a [RFC4122] UUID value is required to be provided in the response header to track the interaction.</param>
        /// <param name="xFapiAuthDate">The time when the customer last logged in to the Data Recipient Software Product. Required for all resource calls (customer present and unattended) if the customer is currently logged. Otherwise optional where the customer has not logged in. Not to be included for unauthenticated calls.</param>
        /// <param name="xFapiCustomerIpAddress">The customer&#39;s original IP address if the customer is currently logged in to the Data Recipient Software Product. The presence of this header indicates that the API is being called in a customer present context. Not to be included for unauthenticated calls.</param>
        /// <param name="xCdsClientHeaders">The customer&#39;s original standard http headers [Base64](#common-field-types) encoded, including the original User Agent header, if the customer is currently logged in to the Data Recipient Software Product. Mandatory for customer present calls.  Not required for unattended or unauthenticated calls.</param>
        /// <response code="200">Success</response>
        /// <response code="0">The following error codes MUST be supported:&lt;br/&gt;&lt;ul class&#x3D;\&quot;error-code-list\&quot;&gt;&lt;li&gt;[400 - Invalid Field](#error-400-field-invalid)&lt;/li&gt;&lt;li&gt;[400 - Invalid Version](#error-400-header-invalid-version)&lt;/li&gt;&lt;li&gt;[406 - Unsupported Version](#error-406-header-unsupported-version)&lt;/li&gt;&lt;/ul&gt;</response>
        [HttpGet]
        [Route("/cds-au/v1/common/customer")]
        [ValidateModelState]
        [SwaggerOperation("GetCustomer")]
        [SwaggerResponse(statusCode: 200, type: typeof(ResponseCommonCustomer), description: "Success")]
        [SwaggerResponse(statusCode: 0, type: typeof(ResponseErrorListV2), description: "The following error codes MUST be supported:&lt;br/&gt;&lt;ul class&#x3D;\&quot;error-code-list\&quot;&gt;&lt;li&gt;[400 - Invalid Field](#error-400-field-invalid)&lt;/li&gt;&lt;li&gt;[400 - Invalid Version](#error-400-header-invalid-version)&lt;/li&gt;&lt;li&gt;[406 - Unsupported Version](#error-406-header-unsupported-version)&lt;/li&gt;&lt;/ul&gt;")]
        public virtual IActionResult GetCustomer([FromHeader][Required()]string xV, [FromHeader]string xMinV, [FromHeader]string xFapiInteractionId, [FromHeader]string xFapiAuthDate, [FromHeader]string xFapiCustomerIpAddress, [FromHeader]string xCdsClientHeaders)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(ResponseCommonCustomer));
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(ResponseErrorListV2));
            string exampleJson = null;
            exampleJson = "{\n  \"data\" : {\n    \"person\" : {\n      \"middleNames\" : [ \"middleNames\", \"middleNames\" ],\n      \"firstName\" : \"firstName\",\n      \"lastName\" : \"lastName\",\n      \"occupationCode\" : \"occupationCode\",\n      \"prefix\" : \"prefix\",\n      \"occupationCodeVersion\" : \"ANZSCO_1220.0_2013_V1.2\",\n      \"suffix\" : \"suffix\",\n      \"lastUpdateTime\" : \"lastUpdateTime\"\n    },\n    \"organisation\" : {\n      \"agentRole\" : \"agentRole\",\n      \"agentLastName\" : \"agentLastName\",\n      \"industryCodeVersion\" : \"ANZSIC_1292.0_2006_V2.0\",\n      \"establishmentDate\" : \"establishmentDate\",\n      \"businessName\" : \"businessName\",\n      \"registeredCountry\" : \"registeredCountry\",\n      \"abn\" : \"abn\",\n      \"acn\" : \"acn\",\n      \"industryCode\" : \"industryCode\",\n      \"organisationType\" : \"COMPANY\",\n      \"legalName\" : \"legalName\",\n      \"isACNCRegistered\" : true,\n      \"agentFirstName\" : \"agentFirstName\",\n      \"shortName\" : \"shortName\",\n      \"lastUpdateTime\" : \"lastUpdateTime\"\n    },\n    \"customerUType\" : \"organisation\"\n  },\n  \"meta\" : \"{}\",\n  \"links\" : {\n    \"self\" : \"self\"\n  }\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<ResponseCommonCustomer>(exampleJson)
            : default(ResponseCommonCustomer);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Get Customer Detail
        /// </summary>
        /// <remarks>Obtain detailed information on the authorised customer within the current session.  Obsolete versions: [v1](includes/obsolete/get-customer-detail-v1.html)</remarks>
        /// <param name="xV">Version of the API end point requested by the client. Must be set to a positive integer. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If the value of [x-min-v](#request-headers) is equal to or higher than the value of [x-v](#request-headers) then the [x-min-v](#request-headers) header should be treated as absent. If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable. See [HTTP Headers](#request-headers)</param>
        /// <param name="xMinV">Minimum version of the API end point requested by the client. Must be set to a positive integer if provided. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable.</param>
        /// <param name="xFapiInteractionId">An [RFC4122](https://tools.ietf.org/html/rfc4122) UUID used as a correlation id. If provided, the data holder must play back this value in the x-fapi-interaction-id response header. If not provided a [RFC4122] UUID value is required to be provided in the response header to track the interaction.</param>
        /// <param name="xFapiAuthDate">The time when the customer last logged in to the Data Recipient Software Product. Required for all resource calls (customer present and unattended) if the customer is currently logged. Otherwise optional where the customer has not logged in. Not to be included for unauthenticated calls.</param>
        /// <param name="xFapiCustomerIpAddress">The customer&#39;s original IP address if the customer is currently logged in to the Data Recipient Software Product. The presence of this header indicates that the API is being called in a customer present context. Not to be included for unauthenticated calls.</param>
        /// <param name="xCdsClientHeaders">The customer&#39;s original standard http headers [Base64](#common-field-types) encoded, including the original User Agent header, if the customer is currently logged in to the Data Recipient Software Product. Mandatory for customer present calls.  Not required for unattended or unauthenticated calls.</param>
        /// <response code="200">Success</response>
        /// <response code="0">The following error codes MUST be supported:&lt;br/&gt;&lt;ul class&#x3D;\&quot;error-code-list\&quot;&gt;&lt;li&gt;[400 - Invalid Field](#error-400-field-invalid)&lt;/li&gt;&lt;li&gt;[400 - Invalid Version](#error-400-header-invalid-version)&lt;/li&gt;&lt;li&gt;[406 - Unsupported Version](#error-406-header-unsupported-version)&lt;/li&gt;&lt;/ul&gt;</response>
        [HttpGet]
        [Route("/cds-au/v1/common/customer/detail")]
        [ValidateModelState]
        [SwaggerOperation("GetCustomerDetail")]
        [SwaggerResponse(statusCode: 200, type: typeof(ResponseCommonCustomerDetailV2), description: "Success")]
        [SwaggerResponse(statusCode: 0, type: typeof(ResponseErrorListV2), description: "The following error codes MUST be supported:&lt;br/&gt;&lt;ul class&#x3D;\&quot;error-code-list\&quot;&gt;&lt;li&gt;[400 - Invalid Field](#error-400-field-invalid)&lt;/li&gt;&lt;li&gt;[400 - Invalid Version](#error-400-header-invalid-version)&lt;/li&gt;&lt;li&gt;[406 - Unsupported Version](#error-406-header-unsupported-version)&lt;/li&gt;&lt;/ul&gt;")]
        public virtual IActionResult GetCustomerDetail([FromHeader][Required()]string xV, [FromHeader]string xMinV, [FromHeader]string xFapiInteractionId, [FromHeader]string xFapiAuthDate, [FromHeader]string xFapiCustomerIpAddress, [FromHeader]string xCdsClientHeaders)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(ResponseCommonCustomerDetailV2));
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(ResponseErrorListV2));
            string exampleJson = null;
            exampleJson = "{\n  \"data\" : {\n    \"customerUType\" : \"organisation\"\n  },\n  \"meta\" : \"{}\",\n  \"links\" : {\n    \"self\" : \"self\"\n  }\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<ResponseCommonCustomerDetailV2>(exampleJson)
            : default(ResponseCommonCustomerDetailV2);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Get Outages
        /// </summary>
        /// <remarks>Obtain a list of scheduled outages for the implementation</remarks>
        /// <param name="xV">Version of the API end point requested by the client. Must be set to a positive integer. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If the value of [x-min-v](#request-headers) is equal to or higher than the value of [x-v](#request-headers) then the [x-min-v](#request-headers) header should be treated as absent. If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable. See [HTTP Headers](#request-headers)</param>
        /// <param name="xMinV">Minimum version of the API end point requested by the client. Must be set to a positive integer if provided. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable.</param>
        /// <response code="200">Success</response>
        /// <response code="0">The following error codes MUST be supported:&lt;br/&gt;&lt;ul class&#x3D;\&quot;error-code-list\&quot;&gt;&lt;li&gt;[400 - Invalid Field](#error-400-field-invalid)&lt;/li&gt;&lt;li&gt;[400 - Invalid Version](#error-400-header-invalid-version)&lt;/li&gt;&lt;li&gt;[406 - Unsupported Version](#error-406-header-unsupported-version)&lt;/li&gt;&lt;/ul&gt;</response>
        [HttpGet]
        [Route("/cds-au/v1/discovery/outages")]
        [ValidateModelState]
        [SwaggerOperation("GetOutages")]
        [SwaggerResponse(statusCode: 200, type: typeof(ResponseDiscoveryOutagesList), description: "Success")]
        [SwaggerResponse(statusCode: 0, type: typeof(ResponseErrorListV2), description: "The following error codes MUST be supported:&lt;br/&gt;&lt;ul class&#x3D;\&quot;error-code-list\&quot;&gt;&lt;li&gt;[400 - Invalid Field](#error-400-field-invalid)&lt;/li&gt;&lt;li&gt;[400 - Invalid Version](#error-400-header-invalid-version)&lt;/li&gt;&lt;li&gt;[406 - Unsupported Version](#error-406-header-unsupported-version)&lt;/li&gt;&lt;/ul&gt;")]
        public virtual IActionResult GetOutages([FromHeader][Required()]string xV, [FromHeader]string xMinV)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(ResponseDiscoveryOutagesList));
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(ResponseErrorListV2));
            string exampleJson = null;
            exampleJson = "{\n  \"data\" : {\n    \"outages\" : [ {\n      \"duration\" : \"duration\",\n      \"outageTime\" : \"outageTime\",\n      \"isPartial\" : true,\n      \"explanation\" : \"explanation\"\n    }, {\n      \"duration\" : \"duration\",\n      \"outageTime\" : \"outageTime\",\n      \"isPartial\" : true,\n      \"explanation\" : \"explanation\"\n    } ]\n  },\n  \"meta\" : \"{}\",\n  \"links\" : {\n    \"self\" : \"self\"\n  }\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<ResponseDiscoveryOutagesList>(exampleJson)
            : default(ResponseDiscoveryOutagesList);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Get Status
        /// </summary>
        /// <remarks>Obtain a health check status for the implementation</remarks>
        /// <param name="xV">Version of the API end point requested by the client. Must be set to a positive integer. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If the value of [x-min-v](#request-headers) is equal to or higher than the value of [x-v](#request-headers) then the [x-min-v](#request-headers) header should be treated as absent. If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable. See [HTTP Headers](#request-headers)</param>
        /// <param name="xMinV">Minimum version of the API end point requested by the client. Must be set to a positive integer if provided. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable.</param>
        /// <response code="200">Success</response>
        /// <response code="0">The following error codes MUST be supported:&lt;br/&gt;&lt;ul class&#x3D;\&quot;error-code-list\&quot;&gt;&lt;li&gt;[400 - Invalid Field](#error-400-field-invalid)&lt;/li&gt;&lt;li&gt;[400 - Invalid Version](#error-400-header-invalid-version)&lt;/li&gt;&lt;li&gt;[406 - Unsupported Version](#error-406-header-unsupported-version)&lt;/li&gt;&lt;/ul&gt;</response>
        [HttpGet]
        [Route("/cds-au/v1/discovery/status")]
        [ValidateModelState]
        [SwaggerOperation("GetStatus")]
        [SwaggerResponse(statusCode: 200, type: typeof(ResponseCommonDiscoveryStatus), description: "Success")]
        [SwaggerResponse(statusCode: 0, type: typeof(ResponseErrorListV2), description: "The following error codes MUST be supported:&lt;br/&gt;&lt;ul class&#x3D;\&quot;error-code-list\&quot;&gt;&lt;li&gt;[400 - Invalid Field](#error-400-field-invalid)&lt;/li&gt;&lt;li&gt;[400 - Invalid Version](#error-400-header-invalid-version)&lt;/li&gt;&lt;li&gt;[406 - Unsupported Version](#error-406-header-unsupported-version)&lt;/li&gt;&lt;/ul&gt;")]
        public virtual IActionResult GetStatus([FromHeader][Required()]string xV, [FromHeader]string xMinV)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(ResponseCommonDiscoveryStatus));
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(ResponseErrorListV2));
            string exampleJson = null;
            exampleJson = "{\n  \"data\" : {\n    \"updateTime\" : \"updateTime\",\n    \"explanation\" : \"explanation\",\n    \"expectedResolutionTime\" : \"expectedResolutionTime\",\n    \"detectionTime\" : \"detectionTime\",\n    \"status\" : \"OK\"\n  },\n  \"meta\" : \"{}\",\n  \"links\" : {\n    \"self\" : \"self\"\n  }\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<ResponseCommonDiscoveryStatus>(exampleJson)
            : default(ResponseCommonDiscoveryStatus);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}