# openapi-generator/aspnetcore/common/src/Agl.Cdr.Common/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "contact@consumerdatastandards.gov.au",
      "name" : "Consumer Data Standards",
      "url" : "https://consumerdatastandards.gov.au"
    },
    "description" : "Consumer Data Standards APIs created by the Data Standards Body (DSB), with the Data Standards Chair as the decision maker to meet the needs of the Consumer Data Right",
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "title" : "CDR Common API",
    "version" : "1.15.0"
  },
  "servers" : [ {
    "url" : "https://data.holder.com.au/cds-au/v1"
  } ],
  "paths" : {
    "/common/customer" : {
      "get" : {
        "description" : "Obtain basic information on the customer that has authorised the current session\n\n### Conventions\nIn the customer payloads relevant conventions are explained here, in one place.\n\n#### Given Names\n\n`firstName` represents the first of a person's given names.\n\n`middleNames` represents a collection of given names if the person has more than one given name.\n\nWhere a data holder holds a person's given names as a single string in source systems, it may not possible in some situations to reliably split these given names into their component first and middle names. In these situations, data holders MAY use the `firstName` field to return the single string of given names and an empty `middleNames` array.\n\nFor example, a person whose given names are \"John Paul Winston\" but the data holder cannot determine what is the first name, can return `\"firstName\": \"John Paul Winston\"`.",
        "operationId" : "getCustomer",
        "parameters" : [ {
          "description" : "Version of the API end point requested by the client. Must be set to a positive integer. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If the value of [x-min-v](#request-headers) is equal to or higher than the value of [x-v](#request-headers) then the [x-min-v](#request-headers) header should be treated as absent. If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable. See [HTTP Headers](#request-headers)",
          "in" : "header",
          "name" : "x-v",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Minimum version of the API end point requested by the client. Must be set to a positive integer if provided. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable.",
          "in" : "header",
          "name" : "x-min-v",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "An [RFC4122](https://tools.ietf.org/html/rfc4122) UUID used as a correlation id. If provided, the data holder must play back this value in the x-fapi-interaction-id response header. If not provided a [RFC4122] UUID value is required to be provided in the response header to track the interaction.",
          "in" : "header",
          "name" : "x-fapi-interaction-id",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "The time when the customer last logged in to the Data Recipient Software Product. Required for all resource calls (customer present and unattended) if the customer is currently logged. Otherwise optional where the customer has not logged in. Not to be included for unauthenticated calls.",
          "in" : "header",
          "name" : "x-fapi-auth-date",
          "schema" : {
            "type" : "string",
            "x-conditional" : true
          },
          "x-conditional" : true
        }, {
          "description" : "The customer's original IP address if the customer is currently logged in to the Data Recipient Software Product. The presence of this header indicates that the API is being called in a customer present context. Not to be included for unauthenticated calls.",
          "in" : "header",
          "name" : "x-fapi-customer-ip-address",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "The customer's original standard http headers [Base64](#common-field-types) encoded, including the original User Agent header, if the customer is currently logged in to the Data Recipient Software Product. Mandatory for customer present calls.  Not required for unattended or unauthenticated calls.",
          "in" : "header",
          "name" : "x-cds-client-headers",
          "schema" : {
            "type" : "string",
            "x-conditional" : true,
            "x-cds-type" : "Base64"
          },
          "x-conditional" : true,
          "x-cds-type" : "Base64"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseCommonCustomer"
                }
              }
            },
            "description" : "Success",
            "headers" : {
              "x-v" : {
                "description" : "The [version](#response-headers) of the API end point that the data holder has responded with.",
                "schema" : {
                  "type" : "string"
                }
              },
              "x-fapi-interaction-id" : {
                "description" : "An [RFC4122](https://tools.ietf.org/html/rfc4122) UUID used as a correlation id. If provided, the data holder must play back this value in the x-fapi-interaction-id response header. If not provided a [RFC4122] UUID value is required to be provided in the response header to track the interaction.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseErrorListV2"
                }
              }
            },
            "description" : "The following error codes MUST be supported:<br/><ul class=\"error-code-list\"><li>[400 - Invalid Field](#error-400-field-invalid)</li><li>[400 - Invalid Version](#error-400-header-invalid-version)</li><li>[406 - Unsupported Version](#error-406-header-unsupported-version)</li></ul>",
            "headers" : {
              "x-fapi-interaction-id" : {
                "description" : "An [RFC4122](https://tools.ietf.org/html/rfc4122) UUID used as a correlation id. If provided, the data holder must play back this value in the x-fapi-interaction-id response header. If not provided a [RFC4122] UUID value is required to be provided in the response header to track the interaction.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Get Customer",
        "tags" : [ "Common", "Customer" ],
        "x-scopes" : [ "common:customer.basic:read" ],
        "x-version" : "1"
      }
    },
    "/common/customer/detail" : {
      "get" : {
        "description" : "Obtain detailed information on the authorised customer within the current session.\n\nObsolete versions: [v1](includes/obsolete/get-customer-detail-v1.html)",
        "operationId" : "getCustomerDetail",
        "parameters" : [ {
          "description" : "Version of the API end point requested by the client. Must be set to a positive integer. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If the value of [x-min-v](#request-headers) is equal to or higher than the value of [x-v](#request-headers) then the [x-min-v](#request-headers) header should be treated as absent. If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable. See [HTTP Headers](#request-headers)",
          "in" : "header",
          "name" : "x-v",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Minimum version of the API end point requested by the client. Must be set to a positive integer if provided. The data holder should respond with the highest supported version between [x-min-v](#request-headers) and [x-v](#request-headers). If all versions requested are not supported then the data holder must respond with a 406 Not Acceptable.",
          "in" : "header",
          "name" : "x-min-v",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "An [RFC4122](https://tools.ietf.org/html/rfc4122) UUID used as a correlation id. If provided, the data holder must play back this value in the x-fapi-interaction-id response header. If not provided a [RFC4122] UUID value is required to be provided in the response header to track the interaction.",
          "in" : "header",
          "name" : "x-fapi-interaction-id",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "The time when the customer last logged in to the Data Recipient Software Product. Required for all resource calls (customer present and unattended) if the customer is currently logged. Otherwise optional where the customer has not logged in. Not to be included for unauthenticated calls.",
          "in" : "header",
          "name" : "x-fapi-auth-date",
          "schema" : {
            "type" : "string",
            "x-conditional" : true
          },
          "x-conditional" : true
        }, {
          "description" : "The customer's original IP address if the customer is currently logged in to the Data Recipient Software Product. The presence of this header indicates that the API is being called in a customer present context. Not to be included for unauthenticated calls.",
          "in" : "header",
          "name" : "x-fapi-customer-ip-address",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "The customer's original standard http headers [Base64](#common-field-types) encoded, including the original User Agent header, if the customer is currently logged in to the Data Recipient Software Product. Mandatory for customer present calls.  Not required for unattended or unauthenticated calls.",
          "in" : "header",
          "name" : "x-cds-client-headers",
          "schema" : {
            "type" : "string",
            "x-conditional" : true,
            "x-cds-type" : "Base64"
          },
          "x-conditional" : true,
          "x-cds-type" : "Base64"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseCommonCustomerDetailV2"
                }
              }
            },
            "description" : "Success",
            "headers" : {
              "x-v" : {
                "description" : "The [version](#response-headers) of the API end point that the data holder has responded with.",
                "schema" : {
                  "type" : "string"
                }
              },
              "x-fapi-interaction-id" : {
                "description" : "An [RFC4122](https://tools.ietf.org/html/rfc4122) UUID used as a correlation id. If provided, the data holder must play back this value in the x-fapi-interaction-id response header. If not provided a [RFC4122] UUID value is required to be provided in the response header to track the interaction.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseErrorListV2"
                }
              }
            },
            "description" : "The following error codes MUST be supported:<br/><ul class=\"error-code-list\"><li>[400 - Invalid Field](#error-400-field-invalid)</li><li>[400 - Invalid Version](#error-400-header-invalid-version)</li><li>[406 - Unsupported Version](#error-406-header-unsupported-version)</li></ul>",
            "headers" : {
              "x-fapi-interaction-id" : {
                "description" : "An [RFC4122](https://tools.ietf.org/html/rfc4122) UUID used as a correlation id. If provided, the data holder must play back this value in the x-fapi-interaction-id response header. If not provided a [RFC4122] UUID value is required to be provided in the response header to track the interaction.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Get Customer Detail",
        "tags" : [ "Common", "Customer" ],
        "x-scopes" : [ "common:customer.detail:read" ],
        "x-version" : "2"
      }
    },
...
  },
  "components" : {
...
    "schemas" : {
...
      "ResponseCommonCustomer" : {
        "example" : {
          "data" : {
            "person" : {
              "middleNames" : [ "middleNames", "middleNames" ],
              "firstName" : "firstName",
              "lastName" : "lastName",
              "occupationCode" : "occupationCode",
              "prefix" : "prefix",
              "occupationCodeVersion" : "ANZSCO_1220.0_2013_V1.2",
              "suffix" : "suffix",
              "lastUpdateTime" : "lastUpdateTime"
            },
            "organisation" : {
              "agentRole" : "agentRole",
              "agentLastName" : "agentLastName",
              "industryCodeVersion" : "ANZSIC_1292.0_2006_V2.0",
              "establishmentDate" : "establishmentDate",
              "businessName" : "businessName",
              "registeredCountry" : "registeredCountry",
              "abn" : "abn",
              "acn" : "acn",
              "industryCode" : "industryCode",
              "organisationType" : "COMPANY",
              "legalName" : "legalName",
              "isACNCRegistered" : true,
              "agentFirstName" : "agentFirstName",
              "shortName" : "shortName",
              "lastUpdateTime" : "lastUpdateTime"
            },
            "customerUType" : "organisation"
          },
          "meta" : "{}",
          "links" : {
            "self" : "self"
          }
        },
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ResponseCommonCustomer_data"
          },
          "links" : {
            "$ref" : "#/components/schemas/Links"
          },
          "meta" : {
            "type" : "object"
          }
        },
        "required" : [ "data", "links" ],
        "type" : "object"
      },
      "ResponseCommonCustomerDetailV2" : {
        "example" : {
          "data" : {
            "customerUType" : "organisation"
          },
          "meta" : "{}",
          "links" : {
            "self" : "self"
          }
        },
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ResponseCommonCustomerDetailV2_data"
          },
          "links" : {
            "$ref" : "#/components/schemas/Links"
          },
          "meta" : {
            "type" : "object"
          }
        },
        "required" : [ "data", "links" ],
        "type" : "object"
      },
...
      "ResponseErrorListV2" : {
        "properties" : {
          "errors" : {
            "description" : "",
            "items" : {
              "$ref" : "#/components/schemas/ResponseErrorListV2_errors"
            },
            "type" : "array"
          }
        },
        "required" : [ "errors" ],
        "type" : "object",
        "x-conditional" : [ "meta" ]
      },
...
    }
  },
...
}